Fix Open-graph meta data image and url
Issue #68. Title and description don't work because the url of the metadata is wrongly configured. Change from site.url -> site.url/page.url

Another small improvement is that when the user uses a featured image in their post, it should be displayed in the sharing link instead of the icon of the site.